+++ CS321_W5D2_BlogAPI/Controllers/BlogsController.cs
@@ -19,8 +19,9 @@ namespace CS321_W5D2_BlogAPI.Controllers
         private readonly IBlogService _blogService;
 
         // TODO: inject BlogService
-        public BlogsController()
+        public BlogsController(IBlogService blogService)
         {
+            _blogService = blogService;
         }
 
         // GET: api/blogs
@@ -30,17 +31,7 @@ namespace CS321_W5D2_BlogAPI.Controllers
         {
             try
             {
-                // TODO: replace the code below with the correct implementation
-                // to return all blogs
-                return Ok(new BlogModel[] {
-                    new BlogModel
-                    {
-                        Id = 1,
-                        Name = "Fix Me!",
-                        Description = "Implement GET /api/blogs",
-                        AuthorName = "unknown",
-                    }
-                });
+                return Ok(_blogService.GetAll().ToApiModels());
             }
             catch (Exception ex)
             {
@@ -56,15 +47,7 @@ namespace CS321_W5D2_BlogAPI.Controllers
         {
             try
             {
-                // TODO: replace the code below with the correct implementation
-                // to return a blog by id
-                return Ok(new BlogModel
-                {
-                    Id = id,
-                    Name = "Fix Me!",
-                    Description = "Implement GET /api/blogs/{id}",
-                    AuthorName = "unknown",
-                });
+                return Ok(_blogService.Get(id).ToApiModel());
             }
             catch (Exception ex)
             {
@@ -94,6 +77,11 @@ namespace CS321_W5D2_BlogAPI.Controllers
         {
             try
             {
+                var existing = _blogService.Get(id);
+
+                if (existing == null)
+                    return NotFound();
+
                 return Ok(_blogService.Update(blog).ToApiModel());
             }
             catch (Exception ex)
@@ -109,6 +97,11 @@ namespace CS321_W5D2_BlogAPI.Controllers
         {
             try
             {
+                var existing = _blogService.Get(id);
+
+                if (existing == null)
+                    return NotFound();
+
                 _blogService.Remove(id);
                 return Ok();
             }
